refactor: split app.go into focused files (config.go, emulator.go)
Split the 1173-line app.go into three focused files:
- app.go (~340 lines): Main(), run(), constants, version handling
- config.go (~670 lines): CLI options, config parsing, system variable init
- emulator.go (~190 lines): embedded emulator helpers, platform detection
